+++ src/wardrobe-share/wardrobe-share.module.ts
@@ -2,12 +2,11 @@ import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Module } from '@nestjs/common';
 import { WardrobeShare } from '../dal/entity/wardrobe-share.entity';
 import { User } from '../dal/entity/user.entity';
-import { AuthModule } from '../auth/auth.module';
 import { WardrobeShareController } from './wardrobe-share.controller';
 import { WardrobeShareService } from './wardrobe-share.service';
 
 @Module({
-  imports: [MikroOrmModule.forFeature([WardrobeShare, User]), AuthModule],
+  imports: [MikroOrmModule.forFeature([WardrobeShare, User])],
   controllers: [WardrobeShareController],
   providers: [WardrobeShareService],
   exports: [WardrobeShareService],